:memo: docs: 경로 변경 및 폴더삭제
